Rename 'Http Sink' to 'Http Listener Sink'
Current implementation of sink creates a server and
listens for clients. It sends the items to the connected
clients. We plan to offer another type of http sink
which creates an http client and posts the items to
an external system. Thus 'Listener' suits current
implementation better.

Fix source consume messages
Allow only POST and PUT for source
Hard-code UTF-8 for default fallback toStringFn