Introduced HTTP(S) Listener Source and WebSocket/SSE sinks (#23)
* Introduced HTTP Listener Source

* Fixed checkstyle, added missing javadocs and readme

* Added module to list of modules

* wip

* Adapt upstream changes regarding JSON parsing

* Small refactor

* Update README

* Added websocket and sse sinks

* remove leftovers and fix broken sentences

* extract listener source context to its own class

* introduce builder for http listener and change portOffset to directly setting the port

* Convert http-sink to use builder

* Refactor http tests to eliminate code duplication

* Add ssl tests for http sink

* add TODOs for javadocs to make the code checkstyle complaint

* add javadoc

* Don't use http listener source while testing http sink, instead use IQueue source

* Rename 'Http Sink' to 'Http Listener Sink'

Current implementation of sink creates a server and
listens for clients. It sends the items to the connected
clients. We plan to offer another type of http sink
which creates an http client and posts the items to
an external system. Thus 'Listener' suits current
implementation better.

Fix source consume messages
Allow only POST and PUT for source
Hard-code UTF-8 for default fallback toStringFn

* fix serialization issue

* Add hostFn to source which enables user to pick an interface

If no hostFn is not provided source binds to all interfaces '0.0.0.0'

Add mutual authentication option to source and sink. if enabled
source/sink will ask for client authentication.

* Add/refactor tests for http module

* Add hostFn and accumulate limit for Http Listener Sinks

Also use ringbuffer to store and retrieve the sink address

* update readme

* address review comments

* address review comments

* add assertion for configured port and path

* use while instead of if in hostFn example of Http connector

* Change sink-name so that it does not include '@' in it

Http sink-name consists of the protocol used (websocket or sseEvent),
'@' character and the port configured, eg 'websocket@8081'. Since
forceTotalParallelismOne uses string partitioning strategy, it tries
to calculate the partition using only the port, the part after the
'@'  character.

* move creating extra member to sink test

Co-authored-by: Emin Demirci <emin@hazelcast.com>
Co-authored-by: Ali Gurbuz <ali@hazelcast.com>