Use ConcurrentLinkedQueue to buffer the messages (#93)
* Use ConcurrentLinkedQueue to buffer the messages

Http Listener sink accumulates the messages in an ArrayList
and sends them when a client connects. But this should be
triggered by a new item from the upstream.

By converting the ArrayList to ConcurrentLinkedQueue, it is
now possible to send the accumulated messages from the
thread which accepts the client connections. With this
change we don't need to wait for a new item from upstream
to trigger sending the accumulated messages.

fixes #89

* address review comments